+++ main.py
@@ -1,41 +1,10 @@
 from contextlib import asynccontextmanager
 from typing import Annotated
-from fastapi import Depends, FastAPI, Form, Request, status
+from fastapi import FastAPI, Form, Request, status
 from fastapi.responses import HTMLResponse, RedirectResponse
 from fastapi.templating import Jinja2Templates
-from pydantic import BaseModel, HttpUrl, PostgresDsn, ValidationError, field_validator
-from sqlmodel import Session, create_engine, SQLModel
-from pydantic_settings import BaseSettings, SettingsConfigDict
-
-class Settings(BaseSettings):
-    postgres_url: PostgresDsn
-
-    @field_validator("postgres_url", mode='after')
-    @classmethod
-    def is_db(cls, postgres_url: PostgresDsn):
-        path = postgres_url.path or ""
-        db_name = path.lstrip("/")
-        if not db_name:
-            raise ValueError("Postgres URL must include a database name")
-        return postgres_url
-
-    model_config = SettingsConfigDict(env_file=".env")
-
-settings = Settings() # type: ignore
-
-engine = create_engine(str(settings.postgres_url))
-
-def create_db_and_tables():
-    SQLModel.metadata.create_all(engine)
-
-# This is a generator which will
-# - 1. return new session with first .next()
-# - 2. close session with second .next()
-def get_session():
-    with Session(engine) as session:
-        yield session
-
-SessionDep = Annotated[Session, Depends(get_session)]
+from pydantic import BaseModel, HttpUrl, ValidationError
+from database import create_db_and_tables
 
 # This is just a function which is essentialy a python context manager but async
 # We can use it like this: async with lifespan(app):
